删除全部Color硬编码，改用从主题属性中读取，后续将支持多主题

+++ app/src/main/java/com/mty/bangcalendar/util/CharacterUtil.kt
@@ -33,7 +33,7 @@ object CharacterUtil {
         val calendarUtilLater = CalendarUtil.dateToCalendarUtil(dateLater)
         val dateEarlier = CalendarUtil.getDate(systemDate.year,systemDate.month,systemDate.day)
         val calendarUtilEarlier = CalendarUtil.dateToCalendarUtil(dateEarlier)
-        return calendarUtilLater - calendarUtilEarlier
+        return (calendarUtilLater - calendarUtilEarlier).toInt()
     }
 
     fun matchCharacter(id: Int) =
